Fix prefix removal for folder paths with symlinks when copying a folder to S3
The local folder argument to a copy or sync to S3 method may be non-standardized (e.g. contain ../) or have symlinks. Removing the folder prefix from listed file paths will not work in such case. This change standardizes and resolves symlinks of both the upload/sync folder argument and the listed file paths in that folder.